+++ test/test_helper.exs
@@ -3,6 +3,7 @@ ExUnit.start(
   exclude: [
     # :ledger,
     # :ledger_log,
+    # :ledger_transaction,
     # :credit_note,
     # :credit_note_log,
     # :event,
@@ -47,6 +48,7 @@ Code.require_file("./test/utils/pix_reversal.exs")
 Code.require_file("./test/utils/pix_chargeback.exs")
 Code.require_file("./test/utils/random.exs")
 Code.require_file("./test/utils/ledger.exs")
+Code.require_file("./test/utils/ledger_transaction.exs")
 Code.require_file("./test/utils/issuing_holder.exs")
 Code.require_file("./test/utils/issuing_withdrawal.exs")
 Code.require_file("./test/utils/issuing_invoice.exs")